+++ nf_core/modules/modules_json.py
@@ -15,7 +15,8 @@ from git.exc import GitCommandError
 from typing_extensions import NotRequired, TypedDict  # for py<3.11
 
 import nf_core.utils
-from nf_core.components.components_utils import NF_CORE_MODULES_NAME, NF_CORE_MODULES_REMOTE, get_components_to_install
+from nf_core.components.components_utils import get_components_to_install
+from nf_core.components.constants import NF_CORE_MODULES_NAME, NF_CORE_MODULES_REMOTE
 from nf_core.modules.modules_repo import ModulesRepo
 from nf_core.pipelines.lint_utils import dump_json_with_prettier
 
@@ -676,7 +677,7 @@ class ModulesJson:
             dump_modules_json = True
             for repo, subworkflows in subworkflows_dict.items():
                 for org, subworkflow in subworkflows:
-                    self.recreate_dependencies(repo, org, subworkflow)
+                    self.recreate_dependencies(repo, org, {"name": subworkflow})
         self.pipeline_components = original_pipeline_components
 
         if dump_modules_json:
@@ -1041,10 +1042,8 @@ class ModulesJson:
         self,
         component_type,
         name,
-        repo_url,
-        install_dir,
         dependent_components,
-    ):
+    ) -> Dict[str, Tuple[str, str, str]]:
         """
         Retrieves all pipeline modules/subworkflows that are reported in the modules.json
         as being installed by the given component
@@ -1052,8 +1051,6 @@ class ModulesJson:
         Args:
             component_type (str): Type of component [modules, subworkflows]
             name (str): Name of the component to find dependencies for
-            repo_url (str): URL of the repository containing the components
-            install_dir (str): Name of the directory where components are installed
 
         Returns:
             (dict[str: str,]): Dictionary indexed with the component names, with component_type as value
@@ -1065,15 +1062,17 @@ class ModulesJson:
         component_types = ["modules"] if component_type == "modules" else ["modules", "subworkflows"]
         # Find all components that have an entry of install by of  a given component, recursively call this function for subworkflows
         for type in component_types:
-            try:
-                components = self.modules_json["repos"][repo_url][type][install_dir].items()
-            except KeyError as e:
-                # This exception will raise when there are only modules installed
-                log.debug(f"Trying to retrieve all {type}. There aren't {type} installed. Failed with error {e}")
-                continue
-            for component_name, component_entry in components:
-                if name in component_entry["installed_by"]:
-                    dependent_components[component_name] = type
+            for repo_url in self.modules_json["repos"].keys():
+                modules_repo = ModulesRepo(repo_url)
+                install_dir = modules_repo.repo_path
+                try:
+                    for comp in self.modules_json["repos"][repo_url][type][install_dir]:
+                        if name in self.modules_json["repos"][repo_url][type][install_dir][comp]["installed_by"]:
+                            dependent_components[comp] = (repo_url, install_dir, type)
+                except KeyError as e:
+                    # This exception will raise when there are only modules installed
+                    log.debug(f"Trying to retrieve all {type}. There aren't {type} installed. Failed with error {e}")
+                    continue
 
         return dependent_components
 
@@ -1261,20 +1260,27 @@ class ModulesJson:
         i.e., no module or subworkflow has been installed by the user in the meantime
         """
 
-        sw_path = Path(self.subworkflows_dir, org, subworkflow)
+        sw_name = subworkflow["name"]
+        sw_path = Path(self.subworkflows_dir, org, sw_name)
         dep_mods, dep_subwfs = get_components_to_install(sw_path)
         assert self.modules_json is not None  # mypy
         for dep_mod in dep_mods:
-            installed_by = self.modules_json["repos"][repo]["modules"][org][dep_mod]["installed_by"]
+            name = dep_mod["name"]
+            current_repo = dep_mod.get("git_remote", repo)
+            current_org = dep_mod.get("org_path", org)
+            installed_by = self.modules_json["repos"][current_repo]["modules"][current_org][name]["installed_by"]
             if installed_by == ["modules"]:
                 self.modules_json["repos"][repo]["modules"][org][dep_mod]["installed_by"] = []
-            if subworkflow not in installed_by:
-                self.modules_json["repos"][repo]["modules"][org][dep_mod]["installed_by"].append(subworkflow)
+            if sw_name not in installed_by:
+                self.modules_json["repos"][repo]["modules"][org][dep_mod]["installed_by"].append(sw_name)
 
         for dep_subwf in dep_subwfs:
-            installed_by = self.modules_json["repos"][repo]["subworkflows"][org][dep_subwf]["installed_by"]
+            name = dep_subwf["name"]
+            current_repo = dep_subwf.get("git_remote", repo)
+            current_org = dep_subwf.get("org_path", org)
+            installed_by = self.modules_json["repos"][current_repo]["subworkflows"][current_org][name]["installed_by"]
             if installed_by == ["subworkflows"]:
                 self.modules_json["repos"][repo]["subworkflows"][org][dep_subwf]["installed_by"] = []
-            if subworkflow not in installed_by:
-                self.modules_json["repos"][repo]["subworkflows"][org][dep_subwf]["installed_by"].append(subworkflow)
+            if sw_name not in installed_by:
+                self.modules_json["repos"][repo]["subworkflows"][org][dep_subwf]["installed_by"].append(sw_name)
             self.recreate_dependencies(repo, org, dep_subwf)